feat: timestamp use u64 instead of u128

+++ src/time.rs
@@ -1,10 +1,10 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
-pub fn now() -> u128 {
+pub fn now() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_millis()
+        .as_millis() as u64
 }
 
 #[cfg(test)]
